Moved GeometricUtils and its tests to "utils" package
