+++ client/src/main/scala/it/cwmp/client/view/game/GameViewActor.scala
@@ -4,9 +4,9 @@ import akka.actor.{Actor, ActorRef, Cancellable}
 import it.cwmp.client.controller.ViewVisibilityMessages.Hide
 import it.cwmp.client.controller.game.GameEngine
 import it.cwmp.client.controller.messages.Initialize
-import it.cwmp.client.model.game.GeometricUtils
 import it.cwmp.client.model.game.distributed.AkkaDistributedState.UpdateState
 import it.cwmp.client.model.game.impl._
+import it.cwmp.client.utils.GeometricUtils
 import it.cwmp.client.view.game.GameViewActor._
 import it.cwmp.client.view.game.model.{CellView, TentacleView}
 import it.cwmp.utils.Logging
