+++ client/src/main/scala/it/cwmp/client/view/game/model/TentacleView.scala
@@ -3,7 +3,8 @@ package it.cwmp.client.view.game.model
 import java.time.{Duration, Instant}
 
 import it.cwmp.client.model.game.impl.{Point, Tentacle}
-import it.cwmp.client.model.game.{GeometricUtils, SizingStrategy}
+import it.cwmp.client.model.game.SizingStrategy
+import it.cwmp.client.utils.GeometricUtils
 import it.cwmp.client.view.game.ColoringStrategy
 import javafx.scene.paint.Color
